requirements-dev.txt
# Optional: staging profiling (set STAGING_PROFILING=True and install these)
-r requirements.txt
django-debug-toolbar==4.2.0
